feat(pages): add ContactPage, HomePage, and NavigationPage classes with form handling and navigation methods refactor(tests): update import paths for HomePage and NavigationPage in test files
