+++ tests/web/ui-components.spec.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "@playwright/test";
-import { HomePage } from "../../pageobject/HomePage";
+import { HomePage } from "../../pages/HomePage";
 
 test.describe("UI Components Tests", () => {
   let homePage: HomePage;
